requirements.txt
flask>=2.0.0
tensorflow>=2.10
pillow
numpy
scikit-learn
matplotlib
gunicorn
flask>=2.0.0
tensorflow>=2.10
pillow
numpy
scikit-learn
matplotlib
